Integrate CustomTransition between ConversationsListViewController and ProfileViewController
