+++ FintechChatApp/ProfileViewController.swift
@@ -41,8 +41,6 @@ class ProfileViewController: UIViewController, IProfileModelDelegate {
         configureParticleEffect()
         initProfileInformation()
         addKeyboardNotifications()
-        
-        transitioningDelegate = self
     }
     
     func applyDependencies(model: IProfileModel, presentationAssembly: IPresentationAssembly) {
@@ -426,78 +424,6 @@ extension ProfileViewController: UITextFieldDelegate, UITextViewDelegate {
     
 }
 
-// MARK: - UIImagePickerControllerDelegate, UINavigationControllerDelegate
-
-extension ProfileViewController: UIImagePickerControllerDelegate, UINavigationControllerDelegate {
-    
-    //Function which present to us ActionSheet (Alert) with a choice of getting image (From Camera or Photo Album).
-    func selectingImage() {
-        
-        //Initialize UIImagePickerController which will be present in one of two types.
-        let picker = UIImagePickerController()
-        picker.allowsEditing = true
-        picker.delegate = self
-        
-        //If Camera or Photo Album not available -> show this alert.
-        let alertWithNotAvailableMessage = UIAlertController(title: "Attention", message: "This type of image sampling is not available on your device.", preferredStyle: .alert)
-        alertWithNotAvailableMessage.addAction(UIAlertAction(title: "Got it", style: .default))
-        
-        //Setup our ActionSheet
-        let actionSheet = UIAlertController(title: "Select image", message: "From one of the options", preferredStyle: .actionSheet)
-        
-        //Selecting from Camera
-        actionSheet.addAction(UIAlertAction(title: "Camera", style: .default, handler: { (_) in
-            if UIImagePickerController.isSourceTypeAvailable(.camera) {
-                picker.sourceType = .camera
-                self.present(picker, animated: true)
-            } else {
-                self.present(alertWithNotAvailableMessage, animated: true)
-            }
-        }))
-        
-        //Selecting from Photo Album.
-        actionSheet.addAction(UIAlertAction(title: "Photo Album", style: .default, handler: { (_) in
-            if UIImagePickerController.isSourceTypeAvailable(.photoLibrary) {
-                picker.sourceType = .photoLibrary
-                self.present(picker, animated: true)
-            } else {
-                self.present(alertWithNotAvailableMessage, animated: true)
-            }
-        }))
-        
-        //Selecting from Server.
-        actionSheet.addAction(UIAlertAction(title: "Load", style: .default, handler: { (_) in
-            guard let imagesVC = self.presentationAssembly?.imagesViewController() else { return }
-            let imagesVCWithNavigation = UINavigationController(rootViewController: imagesVC)
-            imagesVC.delegate = self
-            self.present(imagesVCWithNavigation, animated: true)
-        }))
-        
-        actionSheet.addAction(UIAlertAction(title: "Cancel", style: .cancel, handler: { (_) in
-        }))
-        
-        actionSheet.ignoreNegativeWidthConstraints()
-        self.present(actionSheet, animated: true)
-    }
-    
-    //Picking image from Photo Album or Camera.
-    func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey: Any]) {
-        
-        //Picking...
-        guard let image = info[.editedImage] as? UIImage else {
-            return
-        }
-        
-        //Sending image to our UIImageView.
-        self.profileImageView.image = image
-        picker.dismiss(animated: true)
-        
-        self.saveWithGrandCentralDispatchButton.isEnabled = true
-        self.saveWithOperationsButton.isEnabled = true
-        
-    }
-}
-
 extension ProfileViewController: ImageDelegate {
     func imageTransfer(image: UIImage) {
         
@@ -525,8 +451,11 @@ extension ProfileViewController: UIViewControllerTransitioningDelegate {
     func animationController(forPresented presented: UIViewController,
                              presenting: UIViewController,
                              source: UIViewController) -> UIViewControllerAnimatedTransitioning? {
-        print("an")
-        return CustomTransition(animationDuration: 3.5, animationType: .present)
+        return CustomTransition(animationDuration: 1, animationType: .present)
+    }
+    
+    func animationController(forDismissed dismissed: UIViewController) -> UIViewControllerAnimatedTransitioning? {
+        return CustomTransition(animationDuration: 1, animationType: .dismiss)
     }
     
 }
