+++ FintechChatApp/ProfileViewControllerThemes.swift
@@ -42,6 +42,7 @@ extension ProfileViewController: ThemeableViewController {
         profileDescriptionTextView.backgroundColor = .white
         
         //NavigationBar
+        UINavigationBar.appearance().barTintColor = #colorLiteral(red: 0.9607843137, green: 0.9607843137, blue: 0.9607843137, alpha: 1)
         self.navigationController?.navigationBar.backgroundColor = #colorLiteral(red: 0.9607843137, green: 0.9607843137, blue: 0.9607843137, alpha: 1)
         self.navigationController?.navigationBar.largeTitleTextAttributes = [NSAttributedString.Key.foregroundColor: UIColor.black]
         
@@ -69,6 +70,7 @@ extension ProfileViewController: ThemeableViewController {
         profileDescriptionTextView.backgroundColor = .white
         
         //NavigationBar
+        UINavigationBar.appearance().barTintColor = #colorLiteral(red: 0.9607843137, green: 0.9607843137, blue: 0.9607843137, alpha: 1)
         self.navigationController?.navigationBar.backgroundColor = #colorLiteral(red: 0.9607843137, green: 0.9607843137, blue: 0.9607843137, alpha: 1)
         self.navigationController?.navigationBar.largeTitleTextAttributes = [NSAttributedString.Key.foregroundColor: UIColor.black]
         
@@ -96,6 +98,7 @@ extension ProfileViewController: ThemeableViewController {
         profileDescriptionTextView.backgroundColor = .black
         
         //NavigationBar
+        UINavigationBar.appearance().barTintColor = #colorLiteral(red: 0.2549019754, green: 0.2745098174, blue: 0.3019607961, alpha: 1)
         self.navigationController?.navigationBar.backgroundColor = #colorLiteral(red: 0.2549019754, green: 0.2745098174, blue: 0.3019607961, alpha: 1)
         self.navigationController?.navigationBar.largeTitleTextAttributes = [NSAttributedString.Key.foregroundColor: UIColor.white]
